things

+++ crates/rayexec_execution/src/functions/aggregate/builtin/regr_count.rs
@@ -52,7 +52,7 @@ impl BinaryAggregate for RegrCount {
     fn bind(&self, inputs: Vec<Expression>) -> Result<BindState<Self::BindState>> {
         Ok(BindState {
             state: (),
-            return_type: DataType::Float64,
+            return_type: DataType::Int64,
             inputs,
         })
     }
